+++ request.go
@@ -19,7 +19,7 @@ type Request struct {
 	bodyContent       *[]byte // to cache the request body for multiple reads of ReadEntity
 	pathParameters    map[string]string
 	attributes        map[string]interface{} // for storing request-scoped values
-	selectedRoutePath string // root path + route path that matched the request, e.g. /meetings/{id}/attendees
+	selectedRoutePath string                 // root path + route path that matched the request, e.g. /meetings/{id}/attendees
 }
 
 func newRequest(httpRequest *http.Request) *Request {